Restructure project to use enum providers
Restructure the project to accept enum provider variants as arguments rather
than string names of the provider domains.

In addition, this means that not using structs is possible. This should make the
experience much more ergonomic for the end user.

Lastly, modify the Travis config for faster build times.

Documentation has been added.