A couple minor fixes
Derive Default and Debug from the struct, use is_empty() where possible,
and fix a "memory leak".

As the struct has a new(), may as well derive Default from it (and
Debug, as users should expect that to be available).

Instead of using `vec.len() == 0`, `vec.is_empty()` will be ever so
slightly faster; and easier to read.

Using `concat!()` on each use will give you a `&'static str`, which will
live for the entirity of the program. This isn't needed, as a format
can just be used by itself and does not need a `concat!()`.